Add "OK" and "Cancel" buttons
ColorPickerDialog is now a subclass of AlertDialog, which is recommended
by Android's user interface guidelines. It makes the color picker look a
bit more consistent with Android's other dialog-based preferences.

The "Press on Color to apply" text has been removed now that there are
obvious buttons, but clicking on the color panels will still work.
Clicking on the new color panel is equivalent to clicking "OK", and
clicking on the old color panel is equivalent to clicking "Cancel".